+++ src/tests/test_suite.ts
@@ -5,10 +5,6 @@ import { Clock } from "../infrastructure/clock.js";
 import { TestCaseResult, TestMark, TestMarkValue, TestResult, TestStatus, TestSuiteResult } from "./test_result.js";
 import path from "node:path";
 
-// A simple but full-featured test runner. It allows me to get away from Mocha's idiosyncracies and have
-// more control over test execution, while also shielding me from dependency churn.
-
-
 const DEFAULT_TIMEOUT_IN_MS = 2000;
 
 export interface TestConfig {
@@ -32,31 +28,9 @@ export interface ItOptions {
 
 export type NotifyFn = (testResult: TestCaseResult) => void;
 
-interface Describe {
-	(
-		optionalName?: string | DescribeOptions | DescribeFn,
-		optionalOptions?: DescribeOptions | DescribeFn,
-		describeFn?: DescribeFn
-	): TestSuite,
-	skip: (
-		optionalName?: string | DescribeOptions | DescribeFn,
-		optionalOptions?: DescribeOptions | DescribeFn,
-		descrbeFn?: DescribeFn
-	) => TestSuite,
-	only: (
-		optionalName?: string | DescribeOptions | DescribeFn,
-		optionalOptions?: DescribeOptions | DescribeFn,
-		describeFn?: DescribeFn
-	) => TestSuite,
-}
-type DescribeFn = () => void;
+export type DescribeFn = () => void;
 
-interface It {
-	(name: string, optionalOptions?: ItOptions | ItFn, itFn?: ItFn): void;
-	skip: (name: string, optionalOptions?: ItOptions | ItFn, itFn?: ItFn) => void,
-	only: (name: string, optionalOptions?: ItOptions | ItFn, itFn?: ItFn) => void,
-}
-type ItFn = (testUtilities: TestParameters) => Promise<void> | void;
+export type ItFn = (testUtilities: TestParameters) => Promise<void> | void;
 
 type BeforeAfter = (optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) => void;
 
@@ -88,19 +62,27 @@ interface Runnable {
 	_isSkipped: (mark: TestMarkValue) => boolean,
 }
 
-interface TestContext {
-	describe: Describe,
-	it: It,
+export interface TestContext {
+	describe(
+		optionalName: string | DescribeOptions | DescribeFn | undefined,
+		optionalOptions: DescribeOptions | DescribeFn | undefined,
+		describeFn: DescribeFn | undefined,
+		mark: TestMarkValue,
+	): TestSuite,
+	it(
+		name: string,
+		optionalOptions: ItOptions | ItFn | undefined,
+		itFn: ItFn | undefined,
+		mark: TestMarkValue,
+	): void;
 	beforeAll: BeforeAfter,
 	afterAll: BeforeAfter,
 	beforeEach: BeforeAfter,
 	afterEach: BeforeAfter,
 }
 
-const testContext: TestContext[] = [];
-
 /**
- * A simple but full-featured test runner. It's notable for not using globals.
+ * A simple but full-featured test runner.
  */
 export class TestSuite implements Runnable {
 
@@ -164,25 +146,27 @@ export class TestSuite implements Runnable {
 		}
 	}
 
-	/** @private */
-	static _create(
-		nameOrOptionsOrSuiteFn: string | DescribeOptions | DescribeFn | undefined,
-		optionsOrSuiteFn: DescribeOptions | DescribeFn | undefined,
-		possibleSuiteFn: DescribeFn | undefined,
+	/** Internal use only. */
+	static create(
+		nameOrOptionsOrDescribeFn: string | DescribeOptions | DescribeFn | undefined,
+		optionsOrDescribeFn: DescribeOptions | DescribeFn | undefined,
+		possibleDescribeFn: DescribeFn | undefined,
 		mark: TestMarkValue,
+		testContext: TestContext[],
 	): TestSuite {
 		const DescribeOptionsType = { timeout: Number };
 		ensure.signature(arguments, [
 			[ undefined, DescribeOptionsType, String, Function ],
 			[ undefined, DescribeOptionsType, Function ],
 			[ undefined, Function ],
-			String
+			String,
+			Array,
 		]);
 
-		const { name, options, suiteFn } = decipherOverloadedParameters();
+		const { name, options, fn } = decipherOverloadedParameters();
 
-		if (suiteFn !== undefined) {
-			return this.#runDescribeFunction(suiteFn, name, mark, options.timeout);
+		if (fn !== undefined) {
+			return this.#runDescribeFunction(fn, name, mark, testContext, options.timeout);
 		}
 		else if (mark === TestMark.only) {
 			return new TestSuite(name, mark, {
@@ -196,46 +180,46 @@ export class TestSuite implements Runnable {
 		function decipherOverloadedParameters() {
 			let name: string;
 			let options: DescribeOptions | undefined;
-			let suiteFn: DescribeFn | undefined;
+			let fn: DescribeFn | undefined;
 
-			switch (typeof nameOrOptionsOrSuiteFn) {
+			switch (typeof nameOrOptionsOrDescribeFn) {
 				case "string":
-					name = nameOrOptionsOrSuiteFn;
+					name = nameOrOptionsOrDescribeFn;
 					break;
 				case "object":
-					options = nameOrOptionsOrSuiteFn;
+					options = nameOrOptionsOrDescribeFn;
 					break;
 				case "function":
-					suiteFn = nameOrOptionsOrSuiteFn;
+					fn = nameOrOptionsOrDescribeFn;
 					break;
 				case "undefined":
 					break;
 				default:
-					ensure.unreachable(`Unknown typeof for nameOrOptionsOrSuiteFn: ${typeof nameOrOptionsOrSuiteFn}`);
+					ensure.unreachable(`Unknown typeof for nameOrOptionsOrSuiteFn: ${typeof nameOrOptionsOrDescribeFn}`);
 			}
-			switch (typeof optionsOrSuiteFn) {
+			switch (typeof optionsOrDescribeFn) {
 				case "object":
 					ensure.that(options === undefined, "Received two options parameters");
-					options = optionsOrSuiteFn;
+					options = optionsOrDescribeFn;
 					break;
 				case "function":
-					ensure.that(suiteFn === undefined, "Received two suite function parameters");
-					suiteFn = optionsOrSuiteFn;
+					ensure.that(fn === undefined, "Received two suite function parameters");
+					fn = optionsOrDescribeFn;
 					break;
 				case "undefined":
 					break;
 				default:
-					ensure.unreachable(`Unknown typeof for optionsOrSuiteFn: ${typeof optionsOrSuiteFn}`);
+					ensure.unreachable(`Unknown typeof for optionsOrSuiteFn: ${typeof optionsOrDescribeFn}`);
 			}
-			if (possibleSuiteFn !== undefined) {
-				ensure.that(suiteFn === undefined, "Received two suite function parameters");
-				suiteFn = possibleSuiteFn;
+			if (possibleDescribeFn !== undefined) {
+				ensure.that(fn === undefined, "Received two suite function parameters");
+				fn = possibleDescribeFn;
 			}
 
 			name ??= "";
 			options ??= {};
 
-			return { name, options, suiteFn };
+			return { name, options, fn };
 		}
 	}
 
@@ -243,6 +227,7 @@ export class TestSuite implements Runnable {
 		describeFn: DescribeFn,
 		name: string,
 		mark: TestMarkValue,
+		testContext: TestContext[],
 		timeout?: Milliseconds,
 	): TestSuite {
 		const tests: Runnable[] = [];
@@ -251,31 +236,19 @@ export class TestSuite implements Runnable {
 		const beforeEachFns: BeforeAfterDefinition[] = [];
 		const afterEachFns: BeforeAfterDefinition[] = [];
 
-		const pushTest = <T extends Runnable>(test: T): T => {
-			tests.push(test);
-			return test;
-		};
-
-		const result: Describe = (optionalName, optionalOptions, fn) => this._create(optionalName, optionalOptions, fn, TestMark.none);
-		result.skip = (optionalName, optionalOptions, fn) => this._create(optionalName, optionalOptions, fn, TestMark.skip);
-		result.only = (optionalName, optionalOptions, fn) => this._create(optionalName, optionalOptions, fn, TestMark.only);
-
-		const describe: Describe = (optionalName, optionalOptions, fn) => pushTest(TestSuite._create(optionalName, optionalOptions, fn, TestMark.none));
-		describe.skip = (optionalName, optionalOptions, fn) => pushTest(TestSuite._create(optionalName, optionalOptions, fn, TestMark.skip));
-		describe.only = (optionalName, optionalOptions, fn) => pushTest(TestSuite._create(optionalName, optionalOptions, fn, TestMark.only));
-
-		const it: It = (name, optionalOptions, testCaseFn) => pushTest(new TestCase(name, optionalOptions, testCaseFn, TestMark.none));
-		it.skip = (name, optionalOptions, testCaseFn) => pushTest(new TestCase(name, optionalOptions, testCaseFn, TestMark.skip));
-		it.only = (name, optionalOptions, testCaseFn) => pushTest(new TestCase(name, optionalOptions, testCaseFn, TestMark.only));
-
 		testContext.push({
-			describe,
-			it,
+			describe(optionalName, optionalOptions, fn, mark) {
+				return pushTest(TestSuite.create(optionalName, optionalOptions, fn, mark, testContext));
+			},
+			it(name, optionalOptions, testCaseFn, mark) {
+				pushTest(new TestCase(name, optionalOptions, testCaseFn, mark));
+			},
 			beforeAll: defineBeforeAfterFn(beforeAllFns),
 			afterAll: defineBeforeAfterFn(afterAllFns),
 			beforeEach: defineBeforeAfterFn(beforeEachFns),
 			afterEach: defineBeforeAfterFn(afterEachFns),
 		});
+
 		try {
 			describeFn();
 		}
@@ -284,6 +257,11 @@ export class TestSuite implements Runnable {
 		}
 
 		return new TestSuite(name, mark, { tests, beforeAllFns, afterAllFns, beforeEachFns, afterEachFns, timeout });
+
+		function pushTest<T extends Runnable>(test: T): T {
+			tests.push(test);
+			return test;
+		}
 
 		function defineBeforeAfterFn(beforeAfterArray: BeforeAfterDefinition[]) {
 			return function (optionsOrFnAsync: ItOptions | ItFn, possibleFnAsync?: ItFn) {
@@ -595,164 +573,4 @@ async function runTestFnAsync(
 
 function isSuccess(result: TestCaseResult) {
 	return result.status === TestStatus.pass || result.status === TestStatus.skip;
-}
-
-
-function startTest(
-	optionalName: string | DescribeOptions | DescribeFn | undefined,
-	optionalOptions: DescribeOptions | DescribeFn | undefined,
-	fn: DescribeFn | undefined,
-	mark: TestMarkValue,
-): TestSuite {
-	ensure.that(testContext.length === 0, "test() is not re-entrant [don't run test() inside of test()]");
-
-	try {
-		return TestSuite._create(optionalName, optionalOptions, fn, mark);
-	}
-	finally {
-		ensure.that(testContext.length === 0, "test() didn't clear its context; must be an error in ergotest");
-	}
-}
-
-/**
- * Creates a top-level test suite. In your test module, call this function and `export default` the result. Add `.skip`
- * to skip this test suite and `.only` to only run this test suite.
- * @param {string} [optionalName] The name of the test suite. You can skip this parameter and pass
- *   {@link optionalOptions} or {@link fn} instead.
- * @param {DescribeOptions} [optionalOptions] The test suite options. You can skip this parameter and pass {@link fn}
- *   instead.
- * @param {function} [fn] The body of the test suite. In the body, call {@link describe}, {@link it}, {@link
- *   beforeAll}, {@link afterAll}, {@link beforeEach}, and {@link afterEach} to define the tests in the suite. If
- *   undefined, this test suite will be skipped.
- * @returns {TestSuite} The test suite. You’ll typically `export default` the return value rather than using it
- *   directly.
- */
-export function test(
-	optionalName?: string | DescribeOptions | DescribeFn,
-	optionalOptions?: DescribeOptions | DescribeFn,
-	fn?: DescribeFn,
-): TestSuite {
-	return startTest(optionalName, optionalOptions, fn, TestMark.none);
-}
-
-test.skip = function(
-	optionalName?: string | DescribeOptions | DescribeFn,
-	optionalOptions?: DescribeOptions | DescribeFn,
-	fn?: DescribeFn,
-): TestSuite {
-	return startTest(optionalName, optionalOptions, fn, TestMark.skip);
-};
-
-test.only = function(
-	optionalName?: string | DescribeOptions | DescribeFn,
-	optionalOptions?: DescribeOptions | DescribeFn,
-	fn?: DescribeFn,
-): TestSuite {
-	return startTest(optionalName, optionalOptions, fn, TestMark.only);
-};
-
-/**
- * Adds a nested test suite to the current test suite. Must be run inside of a {@link test} or {@link describe}
- * function. Add `.skip` to skip this test suite and `.only` to only run this test suite.
- * @param {string} [optionalName] The name of the test suite. You can skip this parameter and pass
- *   {@link optionalOptions} or {@link fn} instead.
- * @param {DescribeOptions} [optionalOptions] The test suite options. You can skip this parameter and pass {@link fn}
- *   instead.
- * @param {function} [fn] The body of the test suite. In the body, call {@link describe}, {@link it}, {@link
- *   beforeAll}, {@link afterAll}, {@link beforeEach}, and {@link afterEach} to define the tests in the suite. If
- *   undefined, this test suite will be skipped.
- * @returns {TestSuite} The test suite. You’ll typically ignore the return value.
- */
-export function describe(
-	optionalName?: string | DescribeOptions | DescribeFn,
-	optionalOptions?: DescribeOptions | DescribeFn,
-	fn?: DescribeFn,
-) {
-	currentContext("describe").describe(optionalName, optionalOptions, fn);
-}
-
-describe.skip = function(
-	optionalName?: string | DescribeOptions | DescribeFn,
-	optionalOptions?: DescribeOptions | DescribeFn,
-	fn?: DescribeFn,
-) {
-	currentContext("describe").describe.skip(optionalName, optionalOptions, fn);
-};
-
-describe.only = function(
-	optionalName?: string | DescribeOptions | DescribeFn,
-	optionalOptions?: DescribeOptions | DescribeFn,
-	fn?: DescribeFn,
-) {
-	currentContext("describe").describe.only(optionalName, optionalOptions, fn);
-};
-
-/**
- * Adds a test to the current test suite. Must be run inside of a {@link test} or {@link describe} function. Add
- * `.skip` to skip this test and `.only` to only run this test.
- * @param {string} name The name of the test.
- * @param {ItOptions} [optionalOptions] The test options. You can skip this parameter and pass {@link fnAsync} instead.
- * @param {function} [fnAsync] The body of the test. May be synchronous or asynchronous. If undefined, this test will be
- *   skipped.
- */
-export function it(name: string, optionalOptions?: ItOptions | ItFn, fnAsync?: ItFn) {
-	currentContext("it").it(name, optionalOptions, fnAsync);
-}
-
-it.skip = function it(name: string, optionalOptions?: ItOptions | ItFn, fnAsync?: ItFn) {
-	currentContext("it").it.skip(name, optionalOptions, fnAsync);
-};
-
-it.only = function it(name: string, optionalOptions?: ItOptions | ItFn, fnAsync?: ItFn) {
-	currentContext("it").it.only(name, optionalOptions, fnAsync);
-};
-
-/**
- * Adds a function to run before all the tests in the current test suite. Must be run inside of a {@link test} or
- * {@link describe} function.
- * @param {ItOptions} [optionalOptions] The before/after options. You can skip this parameter and pass @{link fnAsync}
- *   instead.
- * @param {function} fnAsync The function to run. May be synchronous or asynchronous.
- */
-export function beforeAll(optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) {
-	currentContext("beforeAll").beforeAll(optionalOptions, fnAsync);
-}
-
-/**
- * Adds a function to run after all the tests in the current test suite. Must be run inside of a {@link test} or
- * {@link describe} function.
- * @param {ItOptions} [optionalOptions] The before/after options. You can skip this parameter and pass @{link fnAsync}
- *   instead.
- * @param {function} [fnAsync] The function to run. May be synchronous or asynchronous.
- */
-export function afterAll(optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) {
-	currentContext("afterAll").afterAll(optionalOptions, fnAsync);
-}
-
-/**
- * Adds a function to run bfeore each of the tests in the current test suite. Must be run inside of a {@link test} or
- * {@link describe} function.
- * @param {ItOptions} [optionalOptions] The before/after options. You can skip this parameter and pass @{link fnAsync}
- *   instead.
- * @param {function} [fnAsync] The function to run. May be synchronous or asynchronous.
- */
-export function beforeEach(optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) {
-	currentContext("beforeEach").beforeEach(optionalOptions, fnAsync);
-}
-
-/**
- * Adds a function to run after each of the tests in the current test suite. Must be run inside of a {@link test} or
- * {@link describe} function.
- * @param {ItOptions} [optionalOptions] The before/after options. You can skip this parameter and pass @{link fnAsync}
- *   instead.
- * @param {function} [fnAsync] The function to run. May be synchronous or asynchronous.
- */
-export function afterEach(optionalOptions: ItOptions | ItFn, fnAsync?: ItFn) {
-	currentContext("afterEach").afterEach(optionalOptions, fnAsync);
-}
-
-function currentContext(functionName: string) {
-	ensure.that(testContext.length > 0, `${functionName}() must be run inside test()`);
-
-	return testContext[testContext.length - 1];
 }