+++ src/util/_ensure_test.ts
@@ -1,12 +1,12 @@
 // Copyright Titanium I.T. LLC. License granted under terms of "The MIT License."
-import { assert, test, describe, it } from "../tests.js";
+import { assert, describe, it } from "../tests.js";
 import * as ensure from "./ensure.js";
 
 interface NodeError extends Error {
 	stack: string;
 }
 
-export default test(() => {
+export default describe(() => {
 
 	describe("condition checking", () => {