+++ src/util/_type_test.ts
@@ -3,10 +3,10 @@
 /* eslint @typescript-eslint/no-explicit-any: "off" */
 // There are a lot of weird JavaScript manipulations in these tests, so we work around the type-checker with "as any"
 
-import { assert, test, describe, it } from "../tests.js";
+import { assert, describe, it } from "../tests.js";
 import * as type from "./type.js";
 
-export default test(() => {
+export default describe(() => {
 
 	describe("checker", () => {